credit-agreement-app/backend/pyproject.toml
[project]
name = "credit-agreement-backend"
version = "0.1.0"
description = "Credit Agreement Terms Extraction API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.100.0",
    "langchain-google-genai>=2.1.1",
    "langchain-openai>=0.3.10",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "PyPDF2>=3.0.0",
    "python-multipart>=0.0.6",
    "uvicorn>=0.23.0",
]
